fix: bugs
fix lot of bug and layout
1. launch icon
2. disable default account
3. rotating image
4. refresh data (all profile page)
5. total fee calc problem
6. total fee layout double to int
7. move view stub position
8. fix return area not showing
9. add ripple animation

+++ app/src/main/java/com/example/escooter/adapter/RentRecordListAdapter.java
@@ -36,7 +36,7 @@ public class RentRecordListAdapter extends RecyclerView.Adapter<RentRecordListAd
             binding.escooterReturnTime.setText(rentRecord.getEndTime());
             binding.feePerMin.setText(String.valueOf(rentRecord.getFeePerMinutes()));
             binding.duration.setText(String.valueOf(rentRecord.getDuration()));
-            binding.totalFee.setText(String.valueOf(rentRecord.getTotalFee()));
+            binding.totalFee.setText(String.valueOf((int)rentRecord.getTotalFee()));
         }
     }
